#381 reorg etl tests to make process clearer
