+++ src/f10_world/test/test_world_etl2_zoo_agg_to_events_agg.py
@@ -7,175 +7,11 @@ from src.f07_fiscal.fiscal_config import (
     weekday_order_str,
 )
 from src.f08_pidgin.pidgin_config import event_id_str
-from src.f09_brick.pandas_tool import (
-    _get_pidgen_brick_format_filenames,
-    get_sheet_names,
-    upsert_sheet,
-)
+from src.f09_brick.pandas_tool import get_sheet_names, upsert_sheet
 from src.f10_world.world import worldunit_shop
 from src.f10_world.examples.world_env import get_test_worlds_dir, env_dir_setup_cleanup
 from pandas import DataFrame, read_excel as pandas_read_excel
 from os.path import exists as os_path_exists
-
-
-def test_WorldUnit_jungle_to_zoo_staging_CreatesZooFiles(env_dir_setup_cleanup):
-    # ESTABLISH
-    fizz_str = "fizz"
-    fizz_world = worldunit_shop(fizz_str)
-    sue_str = "Sue"
-    event_1 = 1
-    event_2 = 2
-    minute_360 = 360
-    minute_420 = 420
-    hour6am = "6am"
-    hour7am = "7am"
-    ex_file_name = "fizzbuzz.xlsx"
-    jungle_file_path = create_path(fizz_world._jungle_dir, ex_file_name)
-    zoo_file_path = create_path(fizz_world._zoo_dir, "br00003.xlsx")
-    brick_columns = [
-        face_id_str(),
-        event_id_str(),
-        cumlative_minute_str(),
-        fiscal_id_str(),
-        hour_label_str(),
-    ]
-    music23_str = "music23"
-    row1 = [sue_str, event_1, minute_360, music23_str, hour6am]
-    row2 = [sue_str, event_1, minute_420, music23_str, hour7am]
-    row3 = [sue_str, event_2, minute_420, music23_str, hour7am]
-    incomplete_brick_columns = [
-        face_id_str(),
-        event_id_str(),
-        cumlative_minute_str(),
-        fiscal_id_str(),
-    ]
-    incom_row1 = [sue_str, event_1, minute_360, music23_str]
-    incom_row2 = [sue_str, event_1, minute_420, music23_str]
-
-    df1 = DataFrame([row1, row2], columns=brick_columns)
-    df2 = DataFrame([incom_row1, incom_row2], columns=incomplete_brick_columns)
-    df3 = DataFrame([row2, row1, row3], columns=brick_columns)
-    br00003_ex1_str = "example1_br00003"
-    br00003_ex2_str = "example2_br00003"
-    br00003_ex3_str = "example3_br00003"
-    upsert_sheet(jungle_file_path, br00003_ex1_str, df1)
-    upsert_sheet(jungle_file_path, br00003_ex2_str, df2)
-    upsert_sheet(jungle_file_path, br00003_ex3_str, df3)
-    assert os_path_exists(zoo_file_path) is False
-
-    # WHEN
-    fizz_world.jungle_to_zoo_staging()
-
-    # THEN
-    print(f"{zoo_file_path=}")
-    assert os_path_exists(zoo_file_path)
-    x_df = pandas_read_excel(zoo_file_path, sheet_name="zoo_staging")
-    assert set(brick_columns).issubset(set(x_df.columns))
-    file_dir_str = "file_dir"
-    file_name_str = "file_name"
-    sheet_name_str = "sheet_name"
-    assert file_dir_str in set(x_df.columns)
-    assert file_name_str in set(x_df.columns)
-    assert sheet_name_str in set(x_df.columns)
-    assert len(x_df) == 5
-    assert get_sheet_names(zoo_file_path) == ["zoo_staging"]
-
-
-def test_WorldUnit_zoo_staging_to_zoo_agg_CreatesOtxSheets_Scenario0_GroupByWorks(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    fizz_str = "fizz"
-    fizz_world = worldunit_shop(fizz_str)
-    sue_str = "Sue"
-    event_1 = 1
-    minute_360 = 360
-    minute_420 = 420
-    hour6am = "6am"
-    hour7am = "7am"
-    ex_file_name = "fizzbuzz.xlsx"
-    jungle_file_path = create_path(fizz_world._jungle_dir, ex_file_name)
-    zoo_file_path = create_path(fizz_world._zoo_dir, "br00003.xlsx")
-    brick_columns = [
-        face_id_str(),
-        event_id_str(),
-        fiscal_id_str(),
-        hour_label_str(),
-        cumlative_minute_str(),
-    ]
-    music23_str = "music23"
-    row1 = [sue_str, event_1, music23_str, hour6am, minute_360]
-    row2 = [sue_str, event_1, music23_str, hour7am, minute_420]
-    row3 = [sue_str, event_1, music23_str, hour7am, minute_420]
-    df1 = DataFrame([row1, row2, row3], columns=brick_columns)
-    upsert_sheet(jungle_file_path, "example1_br00003", df1)
-    fizz_world.jungle_to_zoo_staging()
-    zoo__staging_df = pandas_read_excel(zoo_file_path, sheet_name="zoo_staging")
-    assert len(zoo__staging_df) == 3
-
-    # WHEN
-    fizz_world.zoo_staging_to_zoo_agg()
-
-    # THEN
-    gen_otx_df = pandas_read_excel(zoo_file_path, sheet_name="zoo_agg")
-    ex_otx_df = DataFrame([row1, row2], columns=brick_columns)
-    print(f"{gen_otx_df.columns=}")
-    assert len(ex_otx_df.columns) == len(gen_otx_df.columns)
-    assert list(ex_otx_df.columns) == list(gen_otx_df.columns)
-    assert len(gen_otx_df) > 0
-    assert len(ex_otx_df) == len(gen_otx_df)
-    assert len(gen_otx_df) == 2
-    assert ex_otx_df.to_csv() == gen_otx_df.to_csv()
-    assert get_sheet_names(zoo_file_path) == ["zoo_staging", "zoo_agg"]
-
-
-def test_WorldUnit_zoo_staging_to_zoo_agg_CreatesOtxSheets_Scenario1_GroupByOnlyNonConflictingRecords(
-    env_dir_setup_cleanup,
-):
-    # ESTABLISH
-    fizz_world = worldunit_shop("fizz")
-    sue_str = "Sue"
-    event_1 = 1
-    minute_360 = 360
-    minute_420 = 420
-    minute_480 = 480
-    hour6am = "6am"
-    hour7am = "7am"
-    ex_file_name = "fizzbuzz.xlsx"
-    jungle_file_path = create_path(fizz_world._jungle_dir, ex_file_name)
-    zoo_file_path = create_path(fizz_world._zoo_dir, "br00003.xlsx")
-    brick_columns = [
-        face_id_str(),
-        event_id_str(),
-        fiscal_id_str(),
-        hour_label_str(),
-        cumlative_minute_str(),
-    ]
-    music23_str = "music23"
-    row1 = [sue_str, event_1, music23_str, hour6am, minute_360]
-    row2 = [sue_str, event_1, music23_str, hour7am, minute_420]
-    row3 = [sue_str, event_1, music23_str, hour7am, minute_480]
-    df1 = DataFrame([row1, row2, row3], columns=brick_columns)
-    upsert_sheet(jungle_file_path, "example1_br00003", df1)
-    fizz_world.jungle_to_zoo_staging()
-    zoo_df = pandas_read_excel(zoo_file_path, sheet_name="zoo_staging")
-    assert len(zoo_df) == 3
-
-    # WHEN
-    fizz_world.zoo_staging_to_zoo_agg()
-
-    # THEN
-    gen_otx_df = pandas_read_excel(zoo_file_path, sheet_name="zoo_agg")
-    ex_otx_df = DataFrame([row1], columns=brick_columns)
-    # print(f"{gen_otx_df.columns=}")
-    print(f"{gen_otx_df=}")
-    assert len(ex_otx_df.columns) == len(gen_otx_df.columns)
-    assert list(ex_otx_df.columns) == list(gen_otx_df.columns)
-    assert len(gen_otx_df) > 0
-    assert len(ex_otx_df) == len(gen_otx_df)
-    assert len(gen_otx_df) == 1
-    assert ex_otx_df.to_csv() == gen_otx_df.to_csv()
-    assert get_sheet_names(zoo_file_path) == ["zoo_staging", "zoo_agg"]
 
 
 def test_WorldUnit_zoo_agg_to_zoo_events_CreatesSheets_Scenario0(
@@ -579,23 +415,3 @@ def test_WorldUnit_set_events_from_events_agg_file_ClearsAttr(env_dir_setup_clea
 
     # THEN
     assert not fizz_world.events
-
-
-def test_get_pidgen_brick_format_filenames_ReturnsObj():
-    # ESTABLISH / WHEN
-    pidgen_brick_filenames = _get_pidgen_brick_format_filenames()
-
-    # THEN
-    print(f"need examples for {pidgen_brick_filenames=}")
-    assert pidgen_brick_filenames == {
-        "br00041.xlsx",
-        "br00042.xlsx",
-        "br00043.xlsx",
-        "br00044.xlsx",
-        "br00045.xlsx",
-        "br00113.xlsx",
-        "br00114.xlsx",
-        "br00115.xlsx",
-        "br00116.xlsx",
-        "br00117.xlsx",
-    }
