+++ src/f10_world/test/test_world_etl5_pidgin_agg_to_faces_pidgin.py
@@ -1,10 +1,5 @@
 from src.f00_instrument.file import create_path
-from src.f04_gift.atom_config import (
-    face_id_str,
-    fiscal_id_str,
-    acct_id_str,
-    owner_id_str,
-)
+from src.f04_gift.atom_config import face_id_str
 from src.f08_pidgin.pidgin_config import (
     event_id_str,
     inx_wall_str,
@@ -21,7 +16,6 @@ from src.f08_pidgin.pidgin_config import (
 )
 from src.f09_brick.pandas_tool import sheet_exists, upsert_sheet
 from src.f10_world.world import worldunit_shop
-
 from src.f10_world.examples.world_env import get_test_worlds_dir, env_dir_setup_cleanup
 from pandas import DataFrame, read_excel as pandas_read_excel
 from pandas.testing import assert_frame_equal as pandas_testing_assert_frame_equal
